Pass in "fetch" to the request functions
The Jest tests operate in a Node environment, which does not have "window.fetch". To get around that, we pass in whatever "fetch" implementation we want when we use it.

Creating a version that does NOT require that is easy enough for users. something like:

  const myFetch = request.fetch.bind(this, window.fetch);

and then everything should work. Hell, we could even provide that implementation...